+++ src/main/java/org/mmarini/wheelly/apis/WithIOFlowable.java
@@ -1,0 +1,9 @@
+package org.mmarini.wheelly.apis;
+
+import io.reactivex.rxjava3.core.Flowable;
+
+public interface WithIOFlowable {
+    Flowable<String> readReadLine();
+
+    Flowable<String> readWriteLine();
+}